:rock: [Boundary_Condition] out-of-line definitions
Problem:
- Boundary conditions are concrete, but are implemented only in header
files which slows recompilation.
- Boundary conditions do not provide access to the bounary ID.

Solution:
- Move the boundary condition functions to an implementation file.
- Add acces to the boundary condition ID.
