+++ cxx/lbm/core/Boundary_Condition.hpp
@@ -11,27 +11,20 @@ namespace lbm::core {
   class Wall final : public JSON_Convertible {
   public:
     Wall() = default;
-    Wall(Boundary_ID boundary)
-        : boundary_{boundary} {}
+    Wall(Boundary_ID boundary);
 
     friend bool
-    operator==(const Wall &wall1, const Wall &wall2) {
-      return wall1.boundary_ == wall2.boundary_;
-    }
+    operator==(const Wall &wall1, const Wall &wall2);
+
+    Boundary_ID
+    boundary() const;
 
   private:
     json
-    get_json() const override {
-      json j = json::object();
-      j["wall"] = json::object();
-      j["wall"]["boundary"] = boundary_;
-      return j;
-    }
+    get_json() const override;
 
     void
-    set_json(const json &j) override {
-      boundary_ = j["wall"]["boundary"];
-    }
+    set_json(const json &j) override;
 
     Boundary_ID boundary_{};
   };
@@ -39,31 +32,23 @@ namespace lbm::core {
   class Symmetry final : public JSON_Convertible {
   public:
     Symmetry() = default;
-    Symmetry(Boundary_ID boundary)
-        : boundary_{boundary} {}
+    Symmetry(Boundary_ID boundary);
 
     friend bool
-    operator==(const Symmetry &symmetry1, const Symmetry &symmetry2) {
-      return symmetry1.boundary_ == symmetry2.boundary_;
-    }
+    operator==(const Symmetry &symmetry1, const Symmetry &symmetry2);
 
     friend bool
-    operator!=(const Symmetry &symmetry1, const Symmetry &symmetry2) {
-      return !(symmetry1 == symmetry2);
-    }
+    operator!=(const Symmetry &symmetry1, const Symmetry &symmetry2);
+
+    Boundary_ID
+    boundary() const;
 
   private:
     json
-    get_json() const override {
-      json j = json::object();
-      j["symmetry"] = boundary_;
-      return j;
-    }
+    get_json() const override;
 
     void
-    set_json(const json &j) override {
-      boundary_ = j["symmetry"];
-    }
+    set_json(const json &j) override;
 
     Boundary_ID boundary_;
   };
@@ -71,34 +56,23 @@ namespace lbm::core {
   class Inlet final : public JSON_Convertible {
   public:
     Inlet() = default;
-    Inlet(Boundary_ID boundary, double inlet_speed)
-        : boundary_{boundary}
-        , inlet_speed_{inlet_speed} {}
+    Inlet(Boundary_ID boundary, double inlet_speed);
 
     friend bool
-    operator==(const Inlet &inlet1, const Inlet &inlet2) {
-      return inlet1.boundary_ == inlet2.boundary_ && inlet1.inlet_speed_ == inlet2.inlet_speed_;
-    }
+    operator==(const Inlet &inlet1, const Inlet &inlet2);
 
     friend bool
-    operator!=(const Inlet &inlet1, const Inlet &inlet2) {
-      return !(inlet1 == inlet2);
-    }
+    operator!=(const Inlet &inlet1, const Inlet &inlet2);
+
+    Boundary_ID
+    boundary() const;
 
   private:
     json
-    get_json() const override {
-      json j = json::object();
-      j["inlet"]["boundary"] = boundary_;
-      j["inlet"]["inletSpeed"] = inlet_speed_;
-      return j;
-    }
+    get_json() const override;
 
     void
-    set_json(const json &j) override {
-      boundary_ = j["inlet"]["boundary"];
-      inlet_speed_ = j["inlet"]["inletSpeed"];
-    }
+    set_json(const json &j) override;
 
     Boundary_ID boundary_{};
     double inlet_speed_{};
@@ -107,35 +81,23 @@ namespace lbm::core {
   class Outlet final : public JSON_Convertible {
   public:
     Outlet() = default;
-    Outlet(Boundary_ID boundary, double outlet_speed)
-        : boundary_{boundary}
-        , outlet_speed_{outlet_speed} {}
+    Outlet(Boundary_ID boundary, double outlet_speed);
 
     friend bool
-    operator==(const Outlet &outlet1, const Outlet &outlet2) {
-      return outlet1.boundary_ == outlet2.boundary_ &&
-             outlet1.outlet_speed_ == outlet2.outlet_speed_;
-    }
+    operator==(const Outlet &outlet1, const Outlet &outlet2);
+
     friend bool
-    operator!=(const Outlet &outlet1, const Outlet &outlet2) {
-      return !(outlet1 == outlet2);
-    }
+    operator!=(const Outlet &outlet1, const Outlet &outlet2);
+
+    Boundary_ID
+    boundary() const;
 
   private:
     json
-    get_json() const override {
-      json j = json::object();
-      j["outlet"] = json::object();
-      j["outlet"]["boundary"] = boundary_;
-      j["outlet"]["outletSpeed"] = outlet_speed_;
-      return j;
-    }
+    get_json() const override;
 
     void
-    set_json(const json &j) override {
-      boundary_ = j["outlet"]["boundary"];
-      outlet_speed_ = j["outlet"]["outletSpeed"];
-    }
+    set_json(const json &j) override;
 
     Boundary_ID boundary_{};
     double outlet_speed_{};
@@ -144,36 +106,23 @@ namespace lbm::core {
   class Pressure_Drop final : public JSON_Convertible {
   public:
     Pressure_Drop() = default;
-    Pressure_Drop(Boundary_ID boundary, double value)
-        : boundary_(boundary)
-        , value_(value) {}
+    Pressure_Drop(Boundary_ID boundary, double value);
 
     friend bool
-    operator==(const Pressure_Drop &pressure_drop1, const Pressure_Drop &pressure_drop2) {
-      return pressure_drop1.boundary_ == pressure_drop2.boundary_ &&
-             pressure_drop1.value_ == pressure_drop2.value_;
-    }
+    operator==(const Pressure_Drop &pressure_drop1, const Pressure_Drop &pressure_drop2);
 
     friend bool
-    operator!=(const Pressure_Drop &pressure_drop1, const Pressure_Drop &pressure_drop2) {
-      return !(pressure_drop1 == pressure_drop2);
-    }
+    operator!=(const Pressure_Drop &pressure_drop1, const Pressure_Drop &pressure_drop2);
+
+    Boundary_ID
+    boundary() const;
 
   private:
     json
-    get_json() const override {
-      json j = json::object();
-      j["pressureDrop"] = json::object();
-      j["pressureDrop"]["boundary"] = boundary_;
-      j["pressureDrop"]["value"] = value_;
-      return j;
-    }
+    get_json() const override;
 
     void
-    set_json(const json &j) override {
-      boundary_ = j["pressureDrop"]["boundary"];
-      value_ = j["pressureDrop"]["value"];
-    }
+    set_json(const json &j) override;
 
     Boundary_ID boundary_{};
     double value_{};
@@ -186,50 +135,23 @@ namespace lbm::core {
 
     Boundary_Condition() = default;
 
-    friend void
-    to_json(json &j, const Boundary_Condition &boundary_condition) {
-      visit([&](const auto &boundary_condition) { j = boundary_condition; }, boundary_condition);
-    }
+    Boundary_ID
+    boundary() const;
 
     friend void
-    from_json(const json &j, Boundary_Condition &boundary_condition) {
-      if (j.contains("wall")) {
-        boundary_condition = Wall(j);
-      } else if (j.contains("symmetry")) {
-        boundary_condition = Symmetry(j);
-      } else if (j.contains("inlet")) {
-        boundary_condition = Inlet(j);
-      } else if (j.contains("outlet")) {
-        boundary_condition = Outlet(j);
-      } else if (j.contains("pressureDrop")) {
-        boundary_condition = Pressure_Drop(j);
-      }
-    }
+    to_json(json &j, const Boundary_Condition &boundary_condition);
+
+    friend void
+    from_json(const json &j, Boundary_Condition &boundary_condition);
 
     friend ostream &
-    operator<<(ostream &os, const Boundary_Condition &boundary_condition) {
-      return os << json(boundary_condition);
-    }
+    operator<<(ostream &os, const Boundary_Condition &boundary_condition);
 
     friend istream &
-    operator>>(istream &is, Boundary_Condition &boundary_condition) {
-      boundary_condition = json::parse(is);
-      return is;
-    }
+    operator>>(istream &is, Boundary_Condition &boundary_condition);
 
     friend bool
-    operator==(const Boundary_Condition &bc1, const Boundary_Condition &bc2) {
-      return visit(
-          []<class U, class V>(const U &bc1, const V &bc2) {
-            if constexpr (same_as<U, V>) {
-              return bc1 == bc2;
-            } else {
-              return false;
-            }
-          },
-          bc1,
-          bc2);
-    }
+    operator==(const Boundary_Condition &bc1, const Boundary_Condition &bc2);
   };
 
   class Boundary_Conditions
@@ -241,19 +163,10 @@ namespace lbm::core {
 
   private:
     json
-    get_json() const override {
-      json j = static_cast<const Base &>(*this);
-      return j;
-    }
+    get_json() const override;
 
     void
-    set_json(const json &j) override {
-      Base::clear();
-      transform(std::begin(j),
-                std::end(j),
-                back_inserter(*this),
-                [](const json &bc) -> Boundary_Condition { return bc; });
-    }
+    set_json(const json &j) override;
   };
 
 } // end of namespace lbm::core
